Major rework
The text size was changed to 11/13 to better match the books I had in
front of me. Garamond has a fairly low x-height, so it makes sense to be
a bit more generous in point size.

This made lots of things fit poorly, which made me consider whether they
really needed to be there. Many did not.